Make all integrer metrics uint types, not float64
A whole bunch of metrics that should be integers are being represented
as floats. I get that we need to emit floats to the DD/prometheus APIs,
but the conversion should be done at the last possible moment.

+++ raingutter/prometheus.go
@@ -39,12 +39,12 @@ var (
 )
 
 func (r *raingutter) recordMetrics(tc *totalConnections, useThreads string) {
-	raingutterActive.WithLabelValues(podName, project, podNameSpace).Observe(r.Active)
-	raingutterQueued.WithLabelValues(podName, project, podNameSpace).Observe(r.Queued)
+	raingutterActive.WithLabelValues(podName, project, podNameSpace).Observe(float64(r.Active))
+	raingutterQueued.WithLabelValues(podName, project, podNameSpace).Observe(float64(r.Queued))
 	if useThreads == "true" {
-		raingutterThreads.WithLabelValues(podName, project, podNameSpace).Set(tc.Count)
+		raingutterThreads.WithLabelValues(podName, project, podNameSpace).Set(float64(tc.Count))
 	} else {
-		raingutterWorkers.WithLabelValues(podName, project, podNameSpace).Set(tc.Count)
+		raingutterWorkers.WithLabelValues(podName, project, podNameSpace).Set(float64(tc.Count))
 	}
 }
